# Precompute norms to speed up Passive-Aggressive