--- a/sklearn/utils/seq_dataset.pxd
+++ b/sklearn/utils/seq_dataset.pxd
@@ -8,15 +8,18 @@ ctypedef np.int32_t INTEGER
 
 cdef class SequentialDataset:
     cdef Py_ssize_t n_samples
+    cdef np.ndarray norms
+    cdef int current_index
 
     cdef void next(self, DOUBLE **x_data_ptr, INTEGER **x_ind_ptr,
                    int *nnz, DOUBLE *y, DOUBLE *sample_weight)
     cdef void shuffle(self, seed)
+    cdef void precompute_norms(self)
+    cdef double get_norm(self)
 
 
 cdef class ArrayDataset(SequentialDataset):
     cdef Py_ssize_t n_features
-    cdef int current_index
     cdef int stride
     cdef DOUBLE *X_data_ptr
     cdef DOUBLE *Y_data_ptr
@@ -29,10 +32,11 @@ cdef class ArrayDataset(SequentialDataset):
     cdef void next(self, DOUBLE **x_data_ptr, INTEGER **x_ind_ptr,
                    int *nnz, DOUBLE *y, DOUBLE *sample_weight)
     cdef void shuffle(self, seed)
+    cdef void precompute_norms(self)
+    cdef double get_norm(self)
 
 
 cdef class CSRDataset(SequentialDataset):
-    cdef int current_index
     cdef int stride
     cdef DOUBLE *X_data_ptr
     cdef INTEGER *X_indptr_ptr
@@ -47,3 +51,5 @@ cdef class CSRDataset(SequentialDataset):
     cdef void next(self, DOUBLE **x_data_ptr, INTEGER **x_ind_ptr,
                    int *nnz, DOUBLE *y, DOUBLE *sample_weight)
     cdef void shuffle(self, seed)
+    cdef void precompute_norms(self)
+    cdef double get_norm(self)