--- a/sklearn/utils/seq_dataset.pyx
+++ b/sklearn/utils/seq_dataset.pyx
@@ -44,6 +44,34 @@ cdef class SequentialDataset:
         """Permutes the ordering of examples.  """
         raise NotImplementedError()
 
+    cdef void precompute_norms(self):
+        cdef int current_index_backup = self.current_index
+        self.current_index = 0
+        cdef DOUBLE * x_data_ptr = NULL
+        cdef INTEGER * x_ind_ptr = NULL
+        cdef int xnnz
+        cdef DOUBLE y = 0.0
+        cdef DOUBLE sample_weight
+
+        cdef np.ndarray[DOUBLE, ndim=1,
+                        mode='c'] norms= np.zeros((self.n_samples,),
+                                                  dtype=np.float64)
+        self.norms = norms
+
+        for _ in range(self.n_samples):
+            self.next( & x_data_ptr, & x_ind_ptr, & xnnz, & y,
+                       & sample_weight)
+            self.norms[self.current_index] = sqnorm(x_data_ptr, x_ind_ptr, xnnz)
+        self.current_index = current_index_backup
+
+    cdef double get_norm(self):
+        """squared norm for the current example.
+
+        self.precompute_norms() must have been called first.
+
+        """
+        return self.norms[self.current_index]
+
 
 cdef class ArrayDataset(SequentialDataset):
     """Dataset backed by a two-dimensional numpy array.
@@ -176,3 +204,13 @@ cdef class CSRDataset(SequentialDataset):
 
     cdef void shuffle(self, seed):
         np.random.RandomState(seed).shuffle(self.index)
+
+
+cdef double sqnorm(DOUBLE * x_data_ptr, INTEGER * x_ind_ptr, int xnnz):
+    cdef double x_norm = 0.0
+    cdef int j
+    cdef double z
+    for j in range(xnnz):
+        z = x_data_ptr[j]
+        x_norm += z * z
+    return x_norm