… sklearn/linear_model/sgd_fast.c
… sklearn/utils/seq_dataset.c