@amueller 2019-06-13
stalled and no speed improvement.